Rename and document HITMARKER_x100000
This marker makes the target unable to Counter or Mirror Coat the damage back to an attacker, however the game does not have a scenario where this is applicable since all its uses are in end-turn events. It's used for all passive damage, where the damage dealt isn't coming from an attacker.
Alternate name: HITMARKER_NO_ATTACKER